Change policy from public read to public write
Add metadata check when update is in question

If the operation is UPDATE, we need to check the metadata to see how the
MD_ProgressCode elements compare.  There can be cases where a doc with a
'complete' code is encountered first, then a doc with an 'ongoing' code
is encountered later.  We don't want to update with the 'ongoing'
document.

The code cannot currently run in production, as it makes a bare requests
call to the object store.