Update method for extracting ARM metadata URL
The existing method worked, but it is better to extract the metadata URL
from the JSON-LD instead of assuming that the metadata URL can be
manufactured from the landing page URL.

Fix tests for updating ojects.  The tests had been marked to skip because
the change involved adding an extra HTTP GET to the member node, which
need to be properly mocked out.

Don't use the namespace map from the document.

Harmonize identifier extraction between ARM, IEDA.  Test both for the
exception being raised and the existence of a log message.

Improve logging message.  The logging message in the situation of  when
one document turns out NOT be be an update of an earlier document but
rather the predecessor document needed improvement.

The XPATH needed to locate the metadata timestamp element needed
improvement.

Refactor test suite.  Mock out the logger for the entire ARM test suite,
it just makes things easier.

Add test for bad XML.

Add more tests for nearly 100% code coverage.