Add regex command line option
The regex option allows the user to restrict to specific documents that
match the regex.  This is useful when there is some sort of network
disruption during the harvest of a single document.  It allows us to get
around the restrictions imposed by the last harvest time.

Simplify ARM, IEDA constructors, just use **kwargs since it all gets
passed to the superclass.