Add tests for dateModified
This is an optional keyword, and SHACL has trouble with it.  SHACL
doesn't want to recognize xsd:date or xsd:datetime as proper datatypes
for validation.  Trying to validate it using regular expressions is
problematic because the regex is not a python regex, but rather an xpath
regex, and the result is something that is not easy for the user to
interpret.  It is easier to validate this as a specific python check.