Shacl tests reinstated
One previous problem was nesting.  Instead of defining a new shape for
each nested property, one should just "nest" it.  That results in a
single error statement instead of a long list of them.

A second problem had to do with sporadic failures.  This is a result of
using sets internally in pyshacl.  If a rule was bad but not
specifically tested, then sporadic failures could happen.  So... no
untested shapes!