Remove the custom logger from shacl validation
Using a custom logger in DEBUG mode is not what we want, because it will
pick up any individual condition that fails.  To see why this is bad,
consider an OR condition, A OR B.  If A fails but B succeeds, the
condition succeeds.  But the DEBUG setting will pick up on A's failure,
and we don't want that.  In INFO mode, the custom logger does not give
us any more information than the report text returned by the pyshacl
validator, so there's no longer any use for the custom logger.