Swap out inline contexts
An inline context will cause a network fetch of the context document
from schema.org, and it will bind the "Date" type to "dateModified".
The endresult is that rdflib does not correctly retrieve the string.
See Redmine issue #8833.

We get around this by swapping out inline contexts with something that
will not trigger the network fetch.