Do not verify SSL connections to sitemaps
For some odd reason, every so often the ARM site issues an SSL
certificate verification error when we try to harvest from them.
Unclear why this is happening, as it does NOT happen with the site
checker (which does everything BUT harvest).  We can get around this by
specifically instructing aiohttp to not verify SSL connections.