Update SO XML validation for non-default cases.
Update the XML validation to work for cases where the format ID is not
the default.  If the default validation fails, bring in the XMLValidator
to find a format ID that works.

Fix problem with the validator where we were getting false positives.
The error traces back to not keeping track of the last format ID that
actually worked.  The tests for this were silently failing because the
tests did not include assertions that there were no log messages at the
ERROR level.

Refactor the process_record method to include retrieval of the landing page URL.

Refactor the XML validator to take the harvester logger as an optional
parameter.  This keeps all log messages with the same logger.

Refactor the core, remove some unused constants.