Add CUASI-specific preprocessing of landing page
If a CUAHSI landing page does not advertise itself as 'PUBLISHED', then
we don't want to bother with it.  This requires extra processing of the
landing page, which none of the other slender nodes require.

Add a SkipError for this purpose, i.e. if a landing page advertises
PUBLIC instead of PUBLISHED, then we want to skip the document.