Add tests
Add test for using a regex with a sitemap. The regex is specified via a
commandline switch with the intention of limiting the number of documents
examined to just those that match the regex.

Add test for SO HTML page missing JSON-LD <SCRIPT>.

Add test for JSON-LD <SCRIPT> without Dataset.  ARM usually has two
JSON-LD <SCRIPT> elements, but only one of them is useful.  In this case,
that element is not present, and we expect a RuntimeError.