Add support for NKN
Do not warn for non-XML headers in case of NKN.  We already know that
the NKN sitemap is an HTML directory listing, so don't muddy the log
with a warning for this.

Change name of site_map attribute to sitemap

Issue SkipError if an NKN metadata.xml doc is not there.  This is pretty
common, so don't count a missing document error in the global error count.

Catch unsupported XML formats early for NKN.