Add support for BCO-DMO
Use the mimetypes module to catch encodingFormats in the SPARQL result
sets.  The encodingFormat key may sometimes have the dataone format ID,
but it could also have a mimetype, which we do not want.  Instead of
hard coding a check for possible mimetypes, we can use the mimetypes
module instead.

Add support for determining flavor of SO content.  We need to determine
between either ARM-style or BCO-DMO style.

Pivot SHACL validation based on flavor of SO content.  If we determine
that the SO content looks like BCO-DMO, then initiate separate SHACL
checks than if we assume the content to be ARM-style.

Use a more generous validation for identifiers.  In the case of ARM, we
have been turning the URLS such as 'https://dx.doi.org/10.5439/1025173'
into just 'doi:10.5439/1025173'.  But for others that do not look like
this, such as BCO-DMO, just accept the URI that is given.

Remove code checking on MD_ProgressCode.  This check was meant to use
additional information in the XML file to see if metadata has been updated
or not.  It could fail for ARM if the progress code was spelled wrong,
which does not get caught in validation.  It meant that their were 3
timestamps to keep track of, and that's really too much.

Give better message if empty landing page content.  There was an ARM
case where a landing page had absolutely no content.  It was erroring out,
but not in a particulary understandable way.

Change MN base URL for bcodmo to "/bcodmo".

Add one mimetype to the list of blessed expected XML headers.

Add new requirements files.  This allows us to install into the miniconda
base environment instead of a new environment.  Remove the old one.