Get BCO-DMO working
Use a more generous validation for identifiers.  In the case of ARM, we
have been turning the URLS such as 'https://dx.doi.org/10.5439/1025173'
into just 'doi:10.5439/1025173'.  But for others that do not look like
this, such as BCO-DMO, just accept the URI that is given.

Introduce determination of the "flavor" of SO content.  ARM and BCO-DMO
are so different that we cannot use the same SHACL files to do that.

Move some tests out of test_so_core that are ARM-specific.