Remove code checking on MD_ProgressCode
This check was meant to use additional information in the XML file to
see if metadata has been updated or not.  It could fail for ARM if the
progress code was spelled wrong, which does not get caught in
validation.  It meant that their were 3 timestamps to keep track of, and
that's really too much.

Give better message if empty landing page content.  There was an ARM
case where a landing page had absolutely no content.  It was erroring out,
but not in a particulary understandable way.